refactor: split functionality according to official fastapi documentation
